+++ nuxt.config.ts
@@ -1,7 +1,4 @@
 export default defineNuxtConfig({
-  future: {
-    compatibilityVersion: 4,
-  },
 
   // Get all the pages, components, composables and plugins from the parent theme
   extends: ['./woonuxt_base'],
@@ -23,6 +20,4 @@ export default defineNuxtConfig({
       failOnError: false,
     },
   },
-
-  compatibilityDate: '2024-07-09',
 });