+++ woonuxt_base/app/components/shopElements/Cart.vue
@@ -4,7 +4,7 @@ const { cart, toggleCart, isUpdatingCart } = useCart();
 
 <template>
   <div class="fixed top-0 bottom-0 right-0 z-50 flex flex-col w-11/12 max-w-lg overflow-x-hidden bg-white shadow-lg">
-    <Icon name="ion:close-outline" class="absolute p-1 bg-white rounded-lg shadow-lg top-6 left-6 md:left-8" size="34" @click="toggleCart(false)" />
+    <Icon name="ion:close-outline" class="absolute p-1 rounded-lg shadow-lg top-6 left-6 md:left-8 cursor-pointer" size="34" @click="toggleCart(false)" />
     <EmptyCart v-if="cart && !cart.isEmpty" class="rounded-lg shadow-lg p-1.5 hover:bg-red-400 hover:text-white" />
 
     <div class="mt-8 text-center">
@@ -18,15 +18,19 @@ const { cart, toggleCart, isUpdatingCart } = useCart();
           <CartCard v-for="item in cart.contents?.nodes" :key="item.key" :item />
         </ul>
         <div class="px-8 mb-8">
-          <NuxtLink class="block p-3 text-lg text-center text-white bg-gray-800 rounded-lg shadow-md justify-evenly hover:bg-gray-900" to="/checkout">
+          <NuxtLink
+            class="block p-3 text-lg text-center text-white bg-gray-800 rounded-lg shadow-md justify-evenly hover:bg-gray-900"
+            to="/checkout"
+            @click.prevent="toggleCart()">
             <span class="mx-2">{{ $t('messages.shop.checkout') }}</span>
             <span v-html="cart.total" />
           </NuxtLink>
         </div>
       </template>
       <!-- Empty Cart Message -->
-      <div v-else class="flex flex-col items-center justify-center flex-1 mb-12">
-        <div class="mb-20 text-xl text-gray-300">{{ $t('messages.shop.cartEmpty') }}</div>
+      <div v-else class="flex flex-col items-center justify-center flex-1 mb-12 text-gray-300">
+        <Icon name="ion:cart-outline" size="96" class="opacity-50 mb-5" />
+        <div class="mb-20 text-xl">{{ $t('messages.shop.cartEmpty') }}</div>
       </div>
     </ClientOnly>
     <!-- Cart Loading Overlay -->