+++ woonuxt_base/app/components/shopElements/PaymentOptions.vue
@@ -33,7 +33,7 @@ onMounted(() => {
       <span class="whitespace-nowrap" v-html="gateway.title" />
       <icon name="ion:checkmark-circle" size="20" class="ml-auto text-primary checkmark opacity-0" />
     </div>
-    <div v-if="activePaymentMethod.description" class="prose">
+    <div v-if="activePaymentMethod.description" class="prose block w-full">
       <p class="text-sm text-gray-500" v-html="activePaymentMethod.description" />
     </div>
   </div>