+++ woonuxt_base/app/components/shopElements/StateSelect.vue
@@ -1,16 +1,15 @@
 <script setup>
-const props = defineProps(['modelValue', 'countryCode']);
+const props = defineProps(['modelValue', 'countryCode', 'countryStates']);
 const emit = defineEmits(['update:modelValue']);
-const states = ref([]);
+const states = ref(props.countryStates || []);
 
 function select(evt) {
   emit('update:modelValue', evt.target.value);
 }
 
 async function updateState() {
-  const country = props?.countryCode || 'IE';
   try {
-    const { countryStates } = await GqlGetStates({ country });
+    const { countryStates } = await GqlGetStates({ country: props?.countryCode || 'IE' });
     states.value = countryStates;
   } catch (error) {
     console.error(error);
@@ -23,23 +22,13 @@ watch(
     updateState();
   },
 );
-
-onMounted(() => {
-  updateState();
-});
 </script>
 
 <template>
-  <select @change="select" v-if="states.length">
+  <select @change="select" v-if="states.length" class="h-[42px]">
     <option v-for="state in states" :key="state.code" :value="state.code" :selected="state.code === props.modelValue">
       {{ state.name }}
     </option>
   </select>
   <input v-else type="text" @change="select" placeholder="State" />
 </template>
-
-<style scoped lang="postcss">
-select {
-  height: 42px;
-}
-</style>